Decompose Call.cs: extract TypeNameResolver, split BuildInvocation
Call.cs mixed two unrelated concerns: deciding the shape of an IL call
site (field/ctor/indexer/property/method) and recursively resolving
Cecil TypeReferences into C# type syntax with generic-parameter
substitution. The latter is the hardest, most bug-prone part of the
file (it's what silently fell back to object in the bug fixed in
4d3f92b) and had zero dependency on call-site shape, so it moves to
its own TypeNameResolver class. BuildInvocation's if/else chain is
split into one named method per branch for readability; the dispatch
itself is unchanged.

Pure refactor: Call's public surface, ContractClassWriter, and
AssemblyAnalyser are untouched. All 49 existing tests pass unmodified.
